allow expects! to handle response types as well as option types, add is-none?
